update: fixed edge case on root warnings
I believe I have it 100% working now. I think there was a strange edge case that maybe had to do with the _layout. I tested it out with the template https://example.routify.dev/ and now the root component in development will not give out the warnings and it will not appear in development.