update: supression working
Solution is kind of tacky but it appears to get the job done a lot better now after coming to the realization that there were some issues where a _Layout.svelte component would warn a <"RootFolderHere"> or <Layout> itself depending on the action the user takes (going to previous page / hmr reload) and it's kind of difficult to figure out what the name should be in order to properly supress the warning. Reason being is that in Line:66 including the replaces might not produce the right output to suppress the warning. In order to figure it out I determine if there are particular attributes that exist in ctx in order to properly figure out what the name value should be. It's kind of difficult to explain but feel free to probe for more questions and also feel free to provide any suggestions if there are any.